feat: add backgrounds. Refactor: start game, over game
