+++ src/common.css
@@ -114,12 +114,10 @@ tbody tr {
 .scroll tbody::-webkit-scrollbar {
     width:2px;
     }
-    /* 滚动槽 */
 .scroll tbody::-webkit-scrollbar-track {
     -webkit-box-shadow:inset 0 0 6px rgba(0,0,0,0.3);
     border-radius:10px;
 }
-    /* 滚动条滑块 */
 .scroll tbody::-webkit-scrollbar-thumb {
     border-radius:10px;
     background:rgba(255,255,255,0.5);
